Ticket System Reword, added Database for statistics and command to view statistics
