+++ src/modules/UtilityHandler.ts
@@ -1,4 +1,4 @@
-import { EmbedBuilder, ChatInputCommandInteraction, Interaction, APIEmbedField, AttachmentBuilder, TextChannel, PermissionFlagsBits, ChannelType, GuildMember, ActionRowBuilder, ButtonBuilder, ButtonStyle } from 'discord.js';
+import { EmbedBuilder, ChatInputCommandInteraction, Interaction, APIEmbedField, AttachmentBuilder, TextChannel } from 'discord.js';
 import Bot from '../Bot';
 import * as config from '../../config.json';
 import { Override } from '../entity/Override';
@@ -701,178 +701,6 @@ export default class UtilityHandler {
             });
             console.error(error);
             return url;
-        }
-    }
-
-    // Ticket System Utilities
-    public async getNextTicketNumber(ticketType: string): Promise<number> {
-        const ticketNumbersPath = path.join(process.cwd(), 'ticket-numbers.json');
-
-        try {
-            const data = await fs.readFile(ticketNumbersPath, 'utf-8');
-            const ticketNumbers = JSON.parse(data);
-
-            // Increment the number for this ticket type
-            ticketNumbers[ticketType] = (ticketNumbers[ticketType] || 0) + 1;
-
-            // Save back to file
-            await fs.writeFile(ticketNumbersPath, JSON.stringify(ticketNumbers, null, 4));
-
-            return ticketNumbers[ticketType];
-        } catch (error) {
-            this.client.logger.error({
-                message: 'Failed to read/write ticket numbers file',
-                error,
-                handler: 'UtilityHandler'
-            });
-
-            // Fallback to 1 if file doesn't exist or is corrupted
-            return 1;
-        }
-    }
-
-    public async createTicketChannel(guild: any, ticketType: string, userId: string, ticketNumber: number): Promise<TextChannel | null> {
-        try {
-            const channelName = `${ticketType}-${ticketNumber.toString().padStart(4, '0')}`;
-
-            // Get admin and owner role IDs
-            const adminRoleId = this.stripRole(this.roles.admin);
-            const ownerRoleId = this.stripRole(this.roles.owner);
-
-            // Create the channel with proper permissions
-            const channel = await guild.channels.create({
-                name: channelName,
-                type: ChannelType.GuildText,
-                parent: null, // No category as requested
-                permissionOverwrites: [
-                    {
-                        id: guild.roles.everyone.id,
-                        deny: [PermissionFlagsBits.ViewChannel]
-                    },
-                    {
-                        id: userId,
-                        allow: [
-                            PermissionFlagsBits.ViewChannel,
-                            PermissionFlagsBits.SendMessages,
-                            PermissionFlagsBits.ReadMessageHistory,
-                            PermissionFlagsBits.AttachFiles,
-                            PermissionFlagsBits.EmbedLinks
-                        ]
-                    },
-                    {
-                        id: adminRoleId,
-                        allow: [
-                            PermissionFlagsBits.ViewChannel,
-                            PermissionFlagsBits.SendMessages,
-                            PermissionFlagsBits.ReadMessageHistory,
-                            PermissionFlagsBits.AttachFiles,
-                            PermissionFlagsBits.EmbedLinks,
-                            PermissionFlagsBits.ManageMessages,
-                            PermissionFlagsBits.ManageChannels
-                        ]
-                    },
-                    {
-                        id: ownerRoleId,
-                        allow: [
-                            PermissionFlagsBits.ViewChannel,
-                            PermissionFlagsBits.SendMessages,
-                            PermissionFlagsBits.ReadMessageHistory,
-                            PermissionFlagsBits.AttachFiles,
-                            PermissionFlagsBits.EmbedLinks,
-                            PermissionFlagsBits.ManageMessages,
-                            PermissionFlagsBits.ManageChannels
-                        ]
-                    }
-                ]
-            });
-
-            this.client.logger.log({
-                message: `Created ticket channel: ${channelName} for user ${userId}`,
-                handler: 'UtilityHandler'
-            }, true);
-
-            return channel;
-        } catch (error) {
-            this.client.logger.error({
-                message: `Failed to create ticket channel for type: ${ticketType}`,
-                error,
-                handler: 'UtilityHandler'
-            });
-            return null;
-        }
-    }
-
-    public async sendTicketWelcomeMessage(channel: TextChannel, userId: string, ticketType: string, formData: any): Promise<void> {
-        try {
-            const adminRole = this.roles.admin;
-            const ownerRole = this.roles.owner;
-
-            // Create welcome message
-            const welcomeMessage = `<@${userId}>, your ticket has been created. An ${adminRole} or ${ownerRole} will be with you shortly.`;
-
-            // Create embed with form data using fields for better organization
-            const embed = new EmbedBuilder()
-                .setTitle(`${this.capitalizeFirstLetter(ticketType)} Ticket`)
-                .setColor(this.colours.lightblue)
-                .setTimestamp()
-                .setAuthor({
-                    name: `User: ${channel.guild.members.cache.get(userId)?.user.username || 'Unknown User'}`,
-                    iconURL: channel.guild.members.cache.get(userId)?.user.displayAvatarURL() || undefined
-                });
-
-            // Format the form data based on ticket type using fields
-            switch (ticketType) {
-                case 'suggestion':
-                    embed.addFields(
-                        { name: 'RSN', value: `\`\`\`${formData.rsn}\`\`\``, inline: false },
-                        { name: 'Suggestion', value: `\`\`\`${formData.suggestion}\`\`\``, inline: false },
-                        { name: 'Why would this work?', value: `\`\`\`${formData.reason}\`\`\``, inline: false }
-                    );
-                    break;
-                case 'report':
-                    embed.addFields(
-                        { name: 'RSN', value: `\`\`\`${formData.rsn}\`\`\``, inline: false },
-                        { name: 'Reported User', value: `\`\`\`${formData.reported_user}\`\`\``, inline: false },
-                        { name: 'Reason', value: `\`\`\`${formData.reason}\`\`\``, inline: false },
-                        { name: 'Description', value: `\`\`\`${formData.description}\`\`\``, inline: false }
-                    );
-                    break;
-                case 'contentcreator':
-                    embed.addFields(
-                        { name: 'RSN', value: `\`\`\`${formData.rsn}\`\`\``, inline: false },
-                        { name: 'Platform URL', value: `\`\`\`${formData.platform_url}\`\`\``, inline: false },
-                        { name: 'Additional Information', value: `\`\`\`${formData.additional}\`\`\``, inline: false }
-                    );
-                    break;
-                case 'other':
-                    embed.addFields(
-                        { name: 'RSN', value: `\`\`\`${formData.rsn}\`\`\``, inline: false },
-                        { name: 'How can we assist?', value: `\`\`\`${formData.assistance}\`\`\``, inline: false }
-                    );
-                    break;
-            }
-
-            // Create close button
-            const closeButton = new ActionRowBuilder<ButtonBuilder>()
-                .addComponents(
-                    new ButtonBuilder()
-                        .setCustomId('ticket_close')
-                        .setLabel('Close')
-                        .setStyle(ButtonStyle.Secondary)
-                );
-
-            await channel.send({ content: welcomeMessage, embeds: [embed], components: [closeButton] });
-
-            this.client.logger.log({
-                message: `Sent welcome message to ticket channel: ${channel.name}`,
-                handler: 'UtilityHandler'
-            }, true);
-        } catch (error) {
-            this.client.logger.error({
-                message: `Failed to send welcome message to ticket channel: ${channel.name}`,
-                error,
-                handler: 'UtilityHandler'
-            });
         }
     }
